ref(dif): Move cache based state management around
